+++ client/src/Styles/Home.module.css
@@ -4,13 +4,17 @@
 
 .container {
   background-image: url("../images/home3.jpg");
-  background-repeat: no-repeat;
-  background-attachment: fixed;
-  background-position: center center;
-  background-size: cover;
-  display: flex;
-  flex-direction: column;
   align-items: center;
+  padding: 0;
+  background-repeat: no-repeat;
+  background-attachment: scroll;
+  background-position: center;
+  background-size: cover;
+  min-height: 100vh;
+}
+
+.nav_bar {
+  margin-top: -1rem;
 }
 
 ::-webkit-scrollbar {
@@ -95,7 +99,7 @@
 
 .imgRecarga {
   position: absolute;
-  margin: -15px 0 0 -70px;
+  margin: -15px 850px;
   border-radius: 2px;
 }
 
@@ -112,4 +116,9 @@
   font-size: 100px;
   margin-top: 22rem;
   margin-left: 2rem;
+}
+
+@media (max-width: 420px) {
+  .container {
+  }
 }
