+++ client/src/components/CountryDetail.jsx
@@ -1,10 +1,10 @@
 import { React, useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
-import { getCountryById /* , clearState */ } from "../../redux/actions";
+import { getCountryById /* , clearState */ } from "../redux/actions";
 import { useDispatch, useSelector } from "react-redux";
 import { Link } from "react-router-dom";
-import Styles from "../../Styles/CountryDetail.module.css";
-import giphy from "../../images/giphy.gif";
+import Styles from "../Styles/CountryDetail.module.css";
+import giphy from "../images/giphy.gif";
 import ActivityCard from "./ActivityCard";
 
 function CountryDetail() {
