+++ client/src/components/SearchBar.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { useState } from "react";
 import { useDispatch } from "react-redux";
-import { searchCountries } from "../../redux/actions";
-import Styles from "../../Styles/SearchBar.module.css";
+import { searchCountries } from "../redux/actions";
+import Styles from "../Styles/SearchBar.module.css";
 
 function SearchBar() {
   const dispatch = useDispatch();
